+++ packages/DouYinRecorder/src/douyin_api.ts
@@ -405,6 +405,7 @@ async function getRoomInfoByMobile(
   } = {},
 ): Promise<RoomInfo> {
   if (!secUserId) {
+    console.error(opts);
     throw new Error("Mobile API need secUserId, please set uid field");
   }
   if (typeof secUserId === "number") {
@@ -424,7 +425,7 @@ async function getRoomInfoByMobile(
     {
       params,
       headers: {
-        cookie: opts.auth,
+        // cookie: opts.auth,
       },
     },
   );