+++ packages/app/src/renderer/src/pages/setting/CommonWebhookSetting.vue
@@ -105,6 +105,13 @@
       style="margin-right: 10px; width: 200px"
       clearable
     />
+    <n-button v-if="data.danmuPreset" text style="margin-right: 10px" @click="openSetting">
+      <template #icon>
+        <n-icon>
+          <SettingsOutline />
+        </n-icon>
+      </template>
+    </n-button>
     <n-button
       v-if="data.danmuPreset && !globalFieldsObj.danmuPreset"
       text
@@ -211,6 +218,18 @@
       style="margin-right: 10px; width: 200px"
       clearable
     />
+    <n-button
+      v-if="props.type === 'global'"
+      text
+      style="margin-right: 10px"
+      @click="navigate('sync')"
+    >
+      <template #icon>
+        <n-icon>
+          <SettingsOutline />
+        </n-icon>
+      </template>
+    </n-button>
     <n-button v-if="data.syncId && !globalFieldsObj.syncId" text @click="data.syncId = null"
       >清除</n-button
     >
@@ -482,7 +501,8 @@
 
   <!-- 非弹幕版相关配置 -->
   <template v-if="data.uid">
-    <n-divider />
+    <div class="divider"></div>
+
     <n-form-item>
       <template #label>
         <Tip
@@ -534,13 +554,17 @@
       >
     </n-form-item>
   </template>
+
+  <DanmuFactorySettingDailog v-model:visible="showDanmuPresetDialog" v-model="danmuPresetModel" />
 </template>
 
 <script setup lang="ts">
+import DanmuFactorySettingDailog from "@renderer/components/DanmuFactorySettingDailog.vue";
 import { useDanmuPreset, useUserInfoStore } from "@renderer/stores";
 import { formatWebhookTitle, formatWebhookPartTitle } from "@renderer/apis/bili";
 import { templateRef } from "@vueuse/core";
 import { uploadTitleTemplate } from "@renderer/enums";
+import { SettingsOutline } from "@vicons/ionicons5";
 
 import type { AppRoomConfig, SyncType } from "@biliLive-tools/types";
 
@@ -561,6 +585,9 @@ const props = defineProps<{
     name: string;
     syncSource: SyncType;
   }[];
+}>();
+const emits = defineEmits<{
+  (e: "navigate", value: string): void;
 }>();
 
 const data = defineModel<AppRoomConfig>("data", {
@@ -590,6 +617,13 @@ const globalFieldsObj = defineModel<{
 const notice = useNotification();
 const { danmuPresetsOptions } = storeToRefs(useDanmuPreset());
 const { userList } = storeToRefs(useUserInfoStore());
+const showDanmuPresetDialog = ref(false);
+const danmuPresetModel = computed({
+  get: () => data.value.danmuPreset || "default",
+  set: (value: string) => {
+    data.value.danmuPreset = value;
+  },
+});
 
 const uploadAfterActionOptions = [
   { label: "无操作", value: "none" },
@@ -727,6 +761,24 @@ const previewPartTitle = async (template: string) => {
 
 const isRoom = computed(() => props.type === "room");
 
+const openSetting = () => {
+  showDanmuPresetDialog.value = true;
+};
+
+watch(
+  () => data.value.danmu,
+  (value) => {
+    if (!value) {
+      if (!globalFieldsObj.value.ffmpegPreset) {
+        data.value.ffmpegPreset = null;
+      }
+      if (!globalFieldsObj.value.danmuPreset) {
+        data.value.danmuPreset = null;
+      }
+    }
+  },
+);
+
 watch(
   () => globalFieldsObj.value,
   () => {
@@ -748,6 +800,10 @@ const previewTitle = async (template: string) => {
     title: data,
     duration: 3000,
   });
+};
+
+const navigate = (routerName: string) => {
+  emits("navigate", routerName);
 };
 </script>
 
@@ -777,5 +833,12 @@ const previewTitle = async (template: string) => {
 h2 {
   margin: 0;
   margin-bottom: 10px;
+}
+
+.divider {
+  height: 1px;
+  background-color: var(--bg-hover);
+  margin-bottom: 14px;
+  margin-top: -10px;
 }
 </style>