+++ packages/DouYinRecorder/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
-- 录制：抖音用户页请求增加额外的错误判断 [#425](https://github.com/renmu123/biliLive-tools/pull/425)
+- 抖音用户页请求增加额外的错误判断 [#425](https://github.com/renmu123/biliLive-tools/pull/425)
+- `mobile`接口不再应用Cookie
 
 # 1.16.0